+++ python_modules/dagster/dagster_tests/cli_tests/test_api_commands.py
@@ -59,13 +59,13 @@ def test_execute_run_with_structured_logs():
         assert "PIPELINE_SUCCESS" in result.stdout, "no match, result: {}".format(result)
 
 
-def runner_execute_step_with_structured_logs(runner, cli_args):
-    result = runner.invoke(api.execute_step_with_structured_logs_command, cli_args)
+def runner_execute_step(runner, cli_args):
+    result = runner.invoke(api.execute_step_command, cli_args)
     if result.exit_code != 0:
         # CliRunner captures stdout so printing it out here
         raise Exception(
             (
-                "dagster runner_execute_step_with_structured_logs commands with cli_args {cli_args} "
+                "dagster runner_execute_step commands with cli_args {cli_args} "
                 'returned exit_code {exit_code} with stdout:\n"{stdout}"'
                 '\n exception: "\n{exception}"'
                 '\n and result as string: "{result}"'
@@ -80,7 +80,7 @@ def runner_execute_step_with_structured_logs(runner, cli_args):
     return result
 
 
-def test_execute_step_with_structured_logs():
+def test_execute_step():
     with get_foo_pipeline_handle() as pipeline_handle:
         runner = CliRunner()
 
@@ -98,16 +98,17 @@ def test_execute_step_with_structured_logs():
                 ExecuteStepArgs(
                     pipeline_origin=pipeline_handle.get_python_origin(),
                     pipeline_run_id=run.run_id,
+                    step_keys_to_execute=None,
                     instance_ref=instance.get_ref(),
                 )
             )
 
-            result = runner_execute_step_with_structured_logs(runner, [input_json],)
+            result = runner_execute_step(runner, [input_json],)
 
         assert "STEP_SUCCESS" in result.stdout
 
 
-def test_execute_step_with_structured_logs_verify_step():
+def test_execute_step_verify_step():
     with get_foo_pipeline_handle() as pipeline_handle:
         runner = CliRunner()
 
@@ -130,6 +131,7 @@ def test_execute_step_with_structured_logs_verify_step():
                 ExecuteStepArgs(
                     pipeline_origin=pipeline_handle.get_python_origin(),
                     pipeline_run_id=run.run_id,
+                    step_keys_to_execute=None,
                     instance_ref=instance.get_ref(),
                 )
             )
@@ -161,7 +163,7 @@ def test_execute_step_with_structured_logs_verify_step():
                     instance, run, retries, step_keys_to_execute=["do_something.compute"]
                 )
 
-            runner_execute_step_with_structured_logs(
+            runner_execute_step(
                 runner, [input_json],
             )