+++ python_modules/libraries/dagster-airflow/dagster_airflow/operators/docker_operator.py
@@ -210,14 +210,11 @@ class DagsterDockerOperator(DockerOperator):
                 pipeline_origin=recon_pipeline.get_python_origin(),
                 pipeline_run_id=self.run_id,
                 instance_ref=self.instance_ref,
-                mode=self.mode,
                 step_keys_to_execute=self.step_keys,
-                run_config=self.run_config,
-                retries_dict={},
             )
         )
 
-        command = "dagster api execute_step_with_structured_logs {}".format(json.dumps(input_json))
+        command = "dagster api execute_step {}".format(json.dumps(input_json))
         self.log.info("Executing: {command}\n".format(command=command))
         return command